Move tests to the identical namespace with main project
