+++ library_test/src/jp/mixi/compatibility/android/text/DateStringCompatTest.java
@@ -1,4 +1,4 @@
-package jp.mixi.compatibility.test;
+package jp.mixi.compatibility.android.text;
 
 import java.util.Date;
